src/test/java/guru/springframework/didemo/controllers/ConstructorInjectedControllerTest.java
/* 
User: Urmi
Date: 12/24/2019 
Time: 6:05 PM
*/

package guru.springframework.didemo.controllers;

import guru.springframework.didemo.services.GreetingServiceImpl;
import org.junit.Before;
import org.junit.Test;

import static org.junit.Assert.assertEquals;

public class ConstructorInjectedControllerTest {

    private ConstructorInjectedController constructorInjectedController;

    @Before
    public void setUp() {
        constructorInjectedController = new ConstructorInjectedController(new GreetingServiceImpl());
    }

    @Test
    public void testGreeting(){
        assertEquals(GreetingServiceImpl.HELLO, constructorInjectedController.sayHello());

    }
}
